requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
frozenlist==1.3.0
httpagentparser==1.9.2
idna==3.3
ipinfo==4.2.0
multidict==6.0.2
PyJWT==1.7.1
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
yarl==1.7.2
